requirements.txt
psycopg2-binary==2.9.6
python-telegram-bot==20.3
python-dotenv==1.0.0
openai==0.28